STCNet/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.4.1=pypi_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.11.0=pypi_0
einops=0.6.1=pypi_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gnutls=3.6.15=he1e5248_0
google-auth=2.38.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.62.3=pypi_0
h5py=3.8.0=pypi_0
idna=3.10=pypi_0
imageio=2.31.2=pypi_0
importlib-metadata=6.7.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.3.2=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.48.0=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
loguru=0.7.3=pypi_0
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.4.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
ninja=1.12.1=h06a4308_0
ninja-base=1.12.1=hdb19cb5_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
opencv-python=4.11.0.86=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
packaging=24.0=pypi_0
pillow=9.3.0=py37hace64e9_1
pip=22.3.1=py37h06a4308_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.1.4=pypi_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=1.8.1=py3.7_cuda11.1_cudnn8.0.5_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.7.3=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.14=h39e8969_0
torchvision=0.9.1=py37_cu111
tqdm=4.67.1=pypi_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=2.0.7=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.6.4=h5eee18b_1
yacs=0.1.8=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.4.9=haebb681_0
